Consolidated the ControlLoop code under the helperclasses folder.
