+++ TeamCode/src/main/java/org/nknsd/teamcode/helperClasses/feedbackcontroller/PidController.java
@@ -1,4 +1,4 @@
-package org.nknsd.teamcode.feedbackcontroller;
+package org.nknsd.teamcode.helperClasses.feedbackcontroller;
 public class PidController implements ControlLoop {
 
     final double pMultiplier;
