+++ TeamCode/src/main/java/org/nknsd/teamcode/programs/tests/WufKillerTester.java
@@ -11,7 +11,7 @@ import org.nknsd.teamcode.components.handlers.ServoHandler;
 import org.nknsd.teamcode.components.handlers.VisionHandler;
 import org.nknsd.teamcode.components.handlers.WufSpotter;
 import org.nknsd.teamcode.components.handlers.statemachine.StateMachine;
-import org.nknsd.teamcode.feedbackcontroller.PidController;
+import org.nknsd.teamcode.helperClasses.feedbackcontroller.PidController;
 import org.nknsd.teamcode.frameworks.NKNComponent;
 import org.nknsd.teamcode.frameworks.NKNProgram;
 import org.nknsd.teamcode.states.WufGrabState;
